==> 1-get_tables_groupstage.ipynb <==
import pandas as pd
from string import ascii_uppercase as alfabeto
import pickle

#bajamos la info de wikipedia
tables=pd.read_html('https://web.archive.org/web/20221115040351/https://en.wikipedia.org/wiki/2022_FIFA_World_Cup')

# A -> H
#12 ->  7 * 8 + 12 = 68

#bajamos datos de las tablas | damos formato a columnas
dict_teams = {}
for letra, i in zip(alfabeto, range(12,68,7)):
    df = tables[i]
    df.rename(columns={df.columns[1]:'Team'},inplace=True)
    df.pop('Qualification')
    dict_teams[f'Group {letra}'] = df   

dict_teams['Group C']

with open('dict_table', 'wb') as output:
    pickle.dump(dict_teams, output)
