==> 2-get_results_and_fixture.ipynb <==
from bs4 import BeautifulSoup
import requests
import pandas as pd

#años en los que se jugaron mundiales
years = [1930,1934,1938,1950,1954,1958,1962,1966,1970,1974,1978,1982,1986,1990,1994,1998,2002,2006,2010,2014,2018]

def get_matches(year,score_=None):
    link = f'https://en.wikipedia.org/wiki/{year}_FIFA_World_Cup'
    response = requests.get(link)
    content = response.text
    soup = BeautifulSoup(content,'lxml')
    matches = soup.find_all('div', class_ = 'footballbox')
    home = []
    score = []
    away = []
    for match in matches:
        home.append(match.find('th',class_ ='fhome').get_text())
        if score_ is None:
            score.append(match.find('th',class_ ='fscore').get_text())
        else:
            score.append('')
        away.append(match.find('th',class_ ='faway').get_text())
    dict_football = {'home':home,'score':score,'away':away}
    df_football = pd.DataFrame(dict_football)
    df_football['year'] = year
    return df_football

#historical data
fifa = [get_matches(year) for year in years]    

df_fifa = pd.concat(fifa,ignore_index = True)

df_fifa.to_csv('fifa_worldcup_historical_data.csv', index=False)

df_fixture = get_matches(2022,'0')
df_fixture.to_csv('fifa_worldcup_fixture.csv', index=False)

df_test = get_matches(2010)
df_test

df_fifa.shape

df_fifa.groupby('year').count()
