==> 3_data_cleaning.ipynb <==
# Limpieza de datos

#librerias
import pandas as pd

### Cargar data

df_historical_data = pd.read_csv('fifa_worldcup_historical_data.csv')#all historical
df_fixture = pd.read_csv("fifa_worldcup_fixture.csv")#matches from 2022
df_missing = pd.read_csv("fifa_worldcup_missing_data.csv")#missing data

df_historical_data.info()

df_fixture['home'] = df_fixture['home'].str.strip()
df_fixture['away'] = df_fixture['away'].str.strip()

### Clean df_missing and add to historical data

#df_missing[df_missing['home'].isnull()]
df_missing.dropna(inplace=True)
df_missing.info()

#add to historical data
df_all_historical =  pd.concat([df_historical_data,df_missing], ignore_index = True)
df_all_historical

#delete duplicates
df_all_historical.drop_duplicates(inplace = True)
df_all_historical.sort_values('year', inplace = True)
df_all_historical

### Clean all historical

delete_index = df_all_historical[(df_all_historical['home'].str.contains('Sweden')) 
                   & df_all_historical['away'].str.contains('Austria') ].index
#eliminamos este registro ya que no nos aporta info

df_all_historical.drop(index=delete_index, inplace=True)

#df_all_historical[df_all_historical['score'].str.contains('[^\d–]')]
df_all_historical['score'] = df_all_historical['score'].str.replace('[^\d–]','',regex = True)
#eliminamos carcateres especiales del marcador

df_all_historical['home'] = df_all_historical['home'].str.strip()
df_all_historical['away'] = df_all_historical['away'].str.strip()
#quitamos posibles espacios que se encuentren

df_all_historical[['HomeGoals', 'AwayGoals']] = df_all_historical['score'].str.split('–', expand = True)
#utilizamos expand para separar el resultado en dos columnas
df_all_historical

df_all_historical.drop('score', axis=1, inplace = True)
df_all_historical

#renameColumns
df_all_historical.rename(columns = {'home':'HomeTeam','away': 'AwayTeam', 'year': 'Year'}, inplace = True)
df_all_historical.head(1)

df_all_historical = df_all_historical.astype({'HomeGoals': int, 'AwayGoals': int})
#convertimos datos a int

df_all_historical['TotalGoals'] = df_all_historical['HomeGoals'] + df_all_historical['AwayGoals']
df_all_historical

### Export clean data

df_all_historical.to_csv('clean_fifa_worldcup_matches.csv', index = False)
df_fixture.to_csv('clean_fifa_worldcup_fixture.csv', index = False)

### Extra verifications

df_all_historical = pd.read_csv('clean_fifa_worldcup_matches.csv')
# verify number of matches per competition
years = [1930, 1934, 1938, 1950, 1954, 1958, 1962, 1966, 1970, 1974,
         1978, 1982, 1986, 1990, 1994, 1998, 2002, 2006, 2010, 2014,
         2018]

for year in years:
    print(year, len(df_all_historical[df_all_historical['Year']==year]))
